+++ src/features/manageProfileSettings/components/ProfileSettings/ProfileSettings.tsx
@@ -11,6 +11,7 @@ import { useLoadingState, useManageProfileSettingsState, useStatusState } from '
 
 import { CustomProfileToggle } from '../CustomProfileToggle';
 import { DefaultProfileOption } from '../DefaultProfileOption';
+import { ProfileSettingsEditor } from '../ProfileSettingsEditor';
 
 import './ProfileSettings.scss';
 
@@ -18,8 +19,9 @@ export const ProfileSettings = () => {
   const { setIsLoading } = useLoadingState();
   const { loadProfile } = useLoadProfile();
   const { loadProfileSettings } = useLoadProfileSettings();
-  const { selectedProfile, setProfileSettings } = useManageProfileSettingsState([
+  const { selectedProfile, setFullProfile, setProfileSettings } = useManageProfileSettingsState([
     'selectedProfile',
+    'setFullProfile',
     'setProfileSettings',
   ]);
   const { addStatusMessagesItem } = useStatusState(['addStatusMessagesItem']);
@@ -30,6 +32,7 @@ export const ProfileSettings = () => {
         try {
           setIsLoading(true);
           const profile = await loadProfile(selectedProfile.id);
+          setFullProfile(profile);
           setProfileSettings(await loadProfileSettings(selectedProfile.id, profile));
         } catch {
           addStatusMessagesItem?.(
@@ -48,12 +51,12 @@ export const ProfileSettings = () => {
     <div data-testid="profile-settings" className="profile-settings">
       <div className="nav">
         <div className="nav-block nav-block-fixed-height">
-          <div className="heading">
+          <h3 className="heading">
             <FormattedMessage
               id={getProfileLabelId(getResourceTypeFromURL(selectedProfile.resourceType as ResourceTypeURL))}
             />
             : {selectedProfile.name}
-          </div>
+          </h3>
           <span className="empty-block" />
         </div>
       </div>
@@ -64,34 +67,7 @@ export const ProfileSettings = () => {
 
       <CustomProfileToggle />
 
-      <div className="settings-sorting">
-        <div className="unused-settings">
-          <h4>
-            <span className="title">
-              <FormattedMessage id="ld.unusedComponents" />
-            </span>
-            (0)
-          </h4>
-          <p>
-            <FormattedMessage id="ld.unusedComponents.description" />
-          </p>
-          <div>
-            <FormattedMessage id="ld.unusedComponents.allUsed" />
-          </div>
-        </div>
-        <div className="active-settings">
-          <h4>
-            <span className="title">
-              <FormattedMessage id="ld.selectedComponents" />
-            </span>
-            (0)
-          </h4>
-          <p>
-            <FormattedMessage id="ld.selectedComponents.description" />
-          </p>
-          <div></div>
-        </div>
-      </div>
+      <ProfileSettingsEditor />
     </div>
   ) : (
     ''