+++ src/store/stores/manageProfileSettings.ts
@@ -1,10 +1,16 @@
+import { DEFAULT_INACTIVE_SETTINGS } from '@/common/constants/profileSettings.constants';
+
 import { type SliceConfigs, createStoreFactory } from '../utils/createStoreFactory';
 import { type SliceState } from '../utils/slice';
 
+type ProfileState = Profile | null;
+type ProfileDTOState = ProfileDTO | null;
+
 export type ManageProfileSettingsState = SliceState<'selectedProfile', ProfileDTO> &
-  SliceState<'nextSelectedProfile', ProfileDTO> &
+  SliceState<'nextSelectedProfile', ProfileDTOState> &
+  SliceState<'fullProfile', ProfileState> &
   SliceState<'isClosingNext', boolean> &
-  SliceState<'profileSettings', ProfileSettings> &
+  SliceState<'profileSettings', ProfileSettingsWithDrift> &
   SliceState<'isTypeDefaultProfile', boolean> &
   SliceState<'isModified', boolean>;
 
@@ -17,13 +23,14 @@ const sliceConfigs: SliceConfigs = {
   nextSelectedProfile: {
     initialValue: null,
   },
+  fullProfile: {
+    initialValue: null,
+  },
   isClosingNext: {
     initialValue: false,
   },
   profileSettings: {
-    initialValue: {
-      active: false,
-    },
+    initialValue: DEFAULT_INACTIVE_SETTINGS,
   },
   isTypeDefaultProfile: {
     initialValue: false,